requirements.txt
pandas
numpy
scikit-learn
seaborn
sentence-transformers
openai 
time
python-dotenv